Move golf-scoring.util.ts into golf module
