+++ backend/src/golf/golf-scoring.util.ts
