+++ backend/src/golf/strategies/golf-teams-game.strategy.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { EventGroup } from '@/events/entities/event-group.entity';
 import { ITeam } from '@/teams/entities/team.entity';
-//TODO move this into the golf module
-import { normalizeName } from '@/player-stats/golf-scoring.util';
+import { normalizeName } from '@/golf/golf-scoring.util';
 import { ITeamsGameStrategy } from '@/teams/strategies/teams-game-strategy.interface';
 
 @Injectable()
